д/з №35 рабочая версия

+++ src/main/resources/static/js/toDoList.js
@@ -1,25 +1,64 @@
 $(document).ready(function () {
-    $('#deleteRow').click(function (){
-        let number;
-        $('.orderNumberHead').each(function (index, value){
-            if($('.orderNumberHead').length <= 1){
-                alert("Удалять нечего!")
-                return;
-            }
-            console.log( index + ": " + value.textContent );
-            let textContent = value.textContent;
-            number=value.textContent
-        })
-        let resultNumber = Number(number);
-        alert(number)
-        $.ajax({
-            url: '/notes/delete/' + resultNumber,
-            type: 'DELETE',
-            success: function () {
-                location.href = '/notes';
-                alert('Удаление прошло успешно!');
-            }
-        });
-    })
+
+    $('.rowInfo > #noteText').blur(function () {
+        let textNote = $(this).text();
+        let textOrder = this.previousElementSibling.textContent;
+        let numberOrder = Number(textOrder);
+        if (textNote === "") {
+            $.ajax({
+                url: '/notes/update/' + numberOrder,
+                type: 'DELETE',
+                success: function () {
+                    alert('Пустая заметка удалена!');
+                    location.href = '/notes';
+                }
+            });
+        } else {
+            $.ajax({
+                url: '/notes/update/' + textNote + '/' + numberOrder,
+                type: 'POST',
+                success: function () {
+                    alert('Заметка обновлена!');
+                }
+            });
+        }
+    })//скриипт обрабатывает текстовое поле заметки
+
+    $('.rowInfo').on('change', '#statusNoteSelector', function () {
+        let val = $(this).val();
+        let parent = $(this).parent().parent().parent();
+        let textOrder = parent[0].firstElementChild.childNodes[0].textContent;
+        let numberNoteOrder = Number(textOrder);
+        // alert(val)
+        // alert(numberNoteOrder)
+        if (val === "0") {
+            let response = String(false);
+            $.ajax({
+                url: '/notes/progress/' + numberNoteOrder + "/" + response,
+                type: 'POST',
+                success: function () {
+                    alert('Заметка в процессе выполнения');
+                }
+            });
+        }
+        if (val === "1") {
+            let response = String(true);
+            $.ajax({
+                url: '/notes/progress/' + numberNoteOrder + "/" + response,
+                type: 'POST',
+                success: function () {
+                    alert('Заметка завершена!');
+                }
+            });
+        }
+    })//скрипт для обработки статуса заметки
+
+    // $('.complete').click(function (){
+    //     $('#onlyOneRow').css('background-color','darkseagreen');
+    // })//скрипт для визуализации выполненной задачи
+    //
+    // $('.inProgress').click(function (){
+    //     $('#onlyOneRow').css('background-color','darkgrey');
+    // })//скрипт для визуализации задачи которая выполняется
 
 });
